pyproject.toml
[tool.poetry]
name = "the-cyber-poet"
version = "0.1.0"
description = ""
authors = ["Christian G <xiangranados@icloud.com>"]
readme = "README.md"
packages = [{include = "the_cyber_poet"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.0"
python-dotenv = "^0.19.2"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
